func + fixes +refactor + docs: dev > main (#3)
# Update .bash_aliases
* docs: 
- Correcting typos.
- Add some comments about the options employed.

* feat: 
- Make cpu/pidstat human-readable, and default to 3 runs
- Stop grepping for known/interesting disks and return all disks, thus making it a more generic alias.
- Show the heading and last part of the path in yellow.
- Add comments to detail used options.
- Add 'personal' choice options to many more alias commands.
- Add shellcheck exceptions.
- Add find file and find directory function.

* fix:
- Correct variable name in 2 locations.
- Correct typos in f_cal() and f_infoExternalIp() that meant they were otherise invalid.
- Remove duplicate findf() definition.
- Removed erroneous # that meant a function was not being created and made the subsequent code syntax incorrect/invalid.
- Add missing ";" to 3 functions.
- fix PERSONAL_LS logic/application.

* refactor: 
-Make one alias reference a second one that does the same thing rather than defining it in full in both places.
- standardise use of 'isPackageInstalled' function

* style: 
- use shfmt (shfmt -i 4 -w filename.sh) on future commits to ensure consistency


# Update .gitconfig
* feat: 
- Added several new defaults


# Update exports
* feat: 
- Add fnm (fast node manager).
- Add a homebrew section for Linux only section.


# Update setup-machine.sh
* docs:
- Adding comments only.

* feat: 
- Add nodejs and related.
- Add gpg package.
- Replace exa with eza (as exa is no longer maintained).
- Introduce cargo 'packages' to the install.
- Add build-essential package to the default installation.
- Add qemu-guest-agent as the default package (as this script gets most used on VM's).
- add pyenv to the standard setup.
- Install rust as part of the machine setup.
- Add oh_my_zsh (zsh manager) to the install routine.
- Add some oh-my-zsh plugins to the base install.
- Only install cargo packages if they are not already installed.
- add fd (a find replacement) to the base install.
- Add go(lang) to default installation.
- golang now only attempts installation if the local version is older than the available version (or is not installed at all).
- Add fzf and shellcheck to the default installation.
- Add gcc and make to default installation.
- Add openssh-client to the base installation.
- Improve function that determines if a package is installed (now using dpkg rather than `command -v` as many packages do not have to use the same internal name and so were falsely reporting not being installed when they were.
- Add more user feedback during the install process.
- add mc (midnight commander) to the default installation.
- Added some feedback to the user on what the script is currently doing.
- Added github configuration.

Inspiration (and some parts copied) from 
https://github.com/schneiderfelipe/dotfiles/blob/f4e1ff1c984eb81bccfa477d8f7f3ca190e63655/up.sh#L40

* fix: 
- Correct use of v(ersion) prefix.
- Use the same checks we introduced to bat.exa installation.
- Add the previously defined install_rust() to the install execution.
- Typo on user output/display.
- Resolve shellcheck.net warning SC2162 (info): read without -r will mangle backslashes.
- Define enable_sshd() which was being called but wasn't defined.
- Corrected go download url.
- Change `curl` to `wget` as the intended file was not getting downloaded.

* refactor:
- Try out a method of not defining an explicit version number for the bat package. Instead, get the latest version from the repository. In addition, check to see if it is already installed and/or with a newer version and interact with the user accordingly. This logic will be carried forward to similar functions in subsequent commits.
- A little bit of standardisation was introduced. (More needs to be done)
- Tidy up apt function.

+++ .bash_aliases
@@ -5,12 +5,33 @@ echo "**** $HOME/.bash_aliases **** starts ****"
 echo ""
 
 # Personalisation
+PERSONAL_ch=true # ch[grp|own|mod]
+PERSONAL_cat=true
 PERSONAL_df=true
+PERSONAL_diff=true
+PERSONAL_grep=true # [e|f]grep
+PERSONAL_ls=true
+PERSONAL_mkdir=true
+PERSONAL_mount=true
+PERSONAL_netstat=true
+PERSONAL_rm=true
+PERSONAL_tree=true
+PERSONAL_wget=true
 
 # The following lines will see your `ls' output colorized
-export LS_OPTIONS='--color=auto'
-eval "$(dircolors)"
-alias ls='ls $LS_OPTIONS --almost-all --classify --human-readable --inode -l --time-style=long-iso'
+if [[ "${PERSONAL_ls}" == "true" ]]; then
+    export LS_OPTIONS='--color=auto'
+
+    if [ -x /usr/bin/dircolors ]; then
+      if test -r "$HOME/.dircolors"; then
+        eval "$(dircolors -b "$HOME/.dircolors")"
+      else
+        eval "$(dircolors -b)"
+      fi
+    fi
+
+    alias ls='ls $LS_OPTIONS --almost-all --classify --human-readable --inode -l --time-style=long-iso'
+fi
 
 # to do: bring the IS_LINUX() under source control so it can be reused here
 # update and upgrade
@@ -25,8 +46,8 @@ alias ls='ls $LS_OPTIONS --almost-all --classify --human-readable --inode -l --t
 
 # Add an "alert" alias for long running commands.  Use like so:
 #   sleep 10; alert
-if command -v notify-send >/dev/null 2>&1; then
-  alias alert='notify-send --urgency=low -i "$([ $? = 0 ] && echo terminal || echo error)" "$(history|tail -n1|sed -e '\''s/^\s*[0-9]\+\s*//;s/[;&|]\s*alert$//'\'')"'
+if isPackageInstalled notify-send; then
+    alias alert='notify-send --urgency=low -i "$([ $? = 0 ] && echo terminal || echo error)" "$(history|tail -n1|sed -e '\''s/^\s*[0-9]\+\s*//;s/[;&|]\s*alert$//'\'')"'
 fi
 
 #  cat options:
@@ -41,20 +62,22 @@ fi
 #    -u                       (ignored)
 #    -v, --show-nonprinting   use ^ and M- notation, except for LFD and TAB
 #
-# There is a namspace issue on Ubuntu, and bat is likely called batcat to avoid it
-if command -v batcat >/dev/null 2>&1; then
-  alias cat='batcat'
-else
-  alias cat='cat --show-all --number'
+# There is a namespace issue on Ubuntu, and bat is likely called batcat to avoid it
+if [[ "${PERSONAL_cat}" == "true" ]]; then
+    if isPackageInstalled batcat; then
+        alias cat='batcat'
+    else
+        alias cat='cat --show-all --number'
+    fi
 fi
 
-alias   ..='cd ..'
+alias ..='cd ..'
 alias cd..='cd ..' # Pickup and fix my common typo
-alias   up='cd ..'
-alias  up2='cd ../..'
-alias  up3='cd ../../..'
+alias up='cd ..'
+alias up2='cd ../..'
+alias up3='cd ../../..'
 
-#to do: bring the IS_WSL() under source control to renenable this
+#to do: bring the IS_WSL() under source control to re-enable this
 #if [[ ${IS_WSL} == true ]]; then
 #    # shellcheck disable=2139
 #    alias cdh="cd /mnt/c/Users/${USER}" # Change to the home directory in Windows
@@ -66,46 +89,61 @@ alias bashfunctions='vi $HOME/.bash_functions && source $HOME/.bash_functions'
 alias calc="bc -l"
 alias countfiles=='find . -type f | wc -l'
 
-alias chgrp='chgrp --preserve-root' # safety net
-alias chmod='chmod --preserve-root' # safeyy net
-alias chown='chown --preserve-root' # safety net
+if [[ "${PERSONAL_ch}" == "true" ]]; then
+    alias chgrp='chgrp --preserve-root' # safety net
+    alias chmod='chmod --preserve-root' # safeyy net
+    alias chown='chown --preserve-root' # safety net
+fi
 
-#to do: make m an n optional parameters of the call (i.e. move this to a function)
+#to do: make m and n optional parameters of the call (i.e. move this to a function)
 # pidstat flags
-# run continously every m seconds (for n times)
-alias cpu='pidstat 5'
+# run continuously every m seconds (for n times)
+alias cpu='pidstat 5 3 --human'
 
 # df options:
 #   -h, --human-readable   : print sizes in powers of 1024 (e.g., 1023M)
-#   -k                     : like --block-size=1K
+#   -l, --local            : limit listing to local file systems
 #   -T, --print-type       : print file system type
-if [[ "${PERSONAL_df}" == "true" ]]
-  then
-        alias df='df -k --human-readable --print-type | grep "disk[^s]\|cache\|shm\|user0\|log\|pool\|Mount" | sort -k 7'
+if [[ "${PERSONAL_df}" == "true" ]]; then
+    # shellcheck disable=SC2142
+    alias df='df --human-readable --local --print-type | head -n 1 | awk '\''{print "\033[33m"$0"\033[39m"}'\''; df --human-readable --local --print-type | tail -n +2 |  sort --version-sort -k 7 | awk '\''BEGIN{FS=OFS="/"} {if (NF>1) {printf "%s", $1; for(i=2; i<NF; i++) printf "/%s", $i; printf "/\033[33m%s\033[39m\n", $NF} else print $0}'\'
 fi
 
-# to do: need to check that colordiff is installed and offer to install it if it isnt (and not set the alias idf it's not)
+# to do: need to check that colordiff is installed and offer to install it if it isn't (and not set the alias if it's not)
 # if it's not installed, alias this instead
 # alias diff='diff --color --side-by-side'
-alias diff="colordiff --color=auto --report-identical-files --side-by-side"
+if [[ "${PERSONAL_diff}" == "true" ]]; then
+    alias diff="colordiff --color=auto --report-identical-files --side-by-side"
+fi
 
-alias dut='df -hl --total | grep total' # total disk usage
+# df options:
+#   -h, --human-readable   : print sizes in powers of 1024 (e.g., 1023M)
+#   -l, --local            : limit listing to local file systems
+#       --total            : elide all entries insignificant to available space,
+#                              and produce a grand total
+# grep options:
+#   -E, --extended-regexp     PATTERNS are extended regular expressions
+alias dut='\df --human-readable --local --total | grep --extended-regexp '"'total|Avail'" # disk usage - total
 
 # egrep option :
 #    --colo[u]r[=WHEN]     : use markers to highlight the matching strings; WHEN is 'always', 'never', or 'auto'
-alias egrep='egrep --color=always'
+if [[ "${PERSONAL_grep}" == "true" ]]; then
+    alias egrep='egrep --color=always'
+fi
 
 # fgrep option :
 #    --colo[u]r[=WHEN]     : use markers to highlight the matching strings; WHEN is 'always', 'never', or 'auto'
-alias fgrep='fgrep --color=always'
+if [[ "${PERSONAL_grep}" == "true" ]]; then
+    alias fgrep='fgrep --color=always'
+fi
 
-if command -v git >/dev/null 2>&1; then
-    alias  ga='gad'
+if isPackageInstalled git; then
+    alias ga='gad'
     alias gad='git add '
     alias gbr='git branch '
     alias gco='git commit'
-    alias  gc='gco'
-    alias  gd='gdi'
+    alias gc='gco'
+    alias gd='gdi'
     alias gdi='git diff'
     alias gch='git checkout '
     alias gita='gad'
@@ -113,17 +151,19 @@ if command -v git >/dev/null 2>&1; then
     alias gitd='gdi'
     alias gitp='gpu'
     alias gits='gst'
-    alias  gp='gpu'
+    alias gp='gpu'
     alias gpu='git pull'
-    alias  gs='gst'
+    alias gs='gst'
     alias gst='git status'
-    alias got='git '          # typos
-    alias gut='git '          # typos
+    alias got='git ' # typos
+    alias gut='git ' # typos
 fi
 
 # grep option :
 #    --colo[u]r[=WHEN]     : use markers to highlight the matching strings; WHEN is 'always', 'never', or 'auto'
-alias grep='grep --color=always'
+if [[ "${PERSONAL_grep}" == "true" ]]; then
+    alias grep='grep --color=always'
+fi
 
 alias h='hist'
 alias hist='history'
@@ -140,28 +180,44 @@ alias infoProcessor='uname --processor'
 alias infoProcessorDetailed='lscpu'
 
 # iostat flags
-# -d m n = run continously every m seconds for n times
+# -d m n = run continuously every m seconds for n times
 # -t   = print the time of each report
 alias io='iostat -d 5 --human -t'
 
-alias kernal='dmesg --color=always --time-format=iso'
+# dmesg options:
+#    -H, --human                 human-readable output
+#    -L, --color[=<when>]        colorize messages (auto, always or never)
+#    -w, --follow                wait for new messages
+#    -W, --follow-new            wait and print only new messages
+#    -x, --decode                decode facility and level to readable string
+#        --time-format <format>  show timestamp using the given format:
+#                                  [delta|reltime|ctime|notime|iso]
+alias kernal='dmesg --color=always --decode --follow --human --time-format=iso'
 
 # mkdir options:
 #    -p, --parents     no error if existing, make parent directories as needed,
 #                      with their file modes unaffected by any -m option.
 #    -v, --verbose     print a message for each created directory
-alias mkdir='mkdir --parents --verbose'
+if [[ "${PERSONAL_mkdir}" == "true" ]]; then
+    alias mkdir='mkdir --parents --verbose'
+fi
 
-if isPackageInstalled ccze
-  then
-    alias mount='mount|sort|column -t|ccze -A' # Make mount readable!
-  else
-    alias mount='mount|sort|column -t' # Make mount readable!
+if [[ "${PERSONAL_mount}" == "true" ]]; then
+    if isPackageInstalled ccze; then
+        alias mount='mount|sort|column -t|ccze -A' # Make mount readable!
+    else
+        alias mount='mount|sort|column -t' # Make mount readable!
+    fi
+fi
 
-alias netstat="netstat -tulanp"
+if [[ "${PERSONAL_netstat}" == "true" ]]; then
+    alias netstat="netstat -tulanp"
+fi
+
 alias now='date +"%d-%m-%Y %T"'
 
 alias path='echo -e ${PATH//:/\\n}'
+# shellcheck disable=SC2142
 alias passwordgenerator='cat /dev/urandom |tr -dc A-Za-z0-9 | head -c${1:-32};echo;'
 
 # rm options:
@@ -174,54 +230,64 @@ alias passwordgenerator='cat /dev/urandom |tr -dc A-Za-z0-9 | head -c${1:-32};ec
 #   --interactive[=WHEN]  : prompt according to WHEN: never, once (-I), or always (-i); without WHEN, prompt always
 #   --preserve-root       : do not remove '/' (default)
 #   --verbose explain what is being done
-alias rm='rm --force --interactive=once --preserve-root --recursive' # safety net
+if [[ "${PERSONAL_rm}" == "true" ]]; then
+    alias rm='rm --force --interactive=once --preserve-root --recursive' # safety net
+fi
 
 alias spacehogs='du -hsx * | sort -rh | head -10'
-alias syslog='tail -f /var/log/syslog'
 
-alias tailKernal='dmesg --color=always --time-format=iso --follow'
+if isPackageInstalled ccze; then
+    alias syslog='tail -f /var/log/syslog | ccze'
+else
+    alias syslog='tail -f /var/log/syslog'
+fi
 
-if command -v tmux >/dev/null 2>&1; then
-  alias tmux='tmux -2'
+alias tailKernal='kernal'
+
+if isPackageInstalled tmux; then
+    alias tmux='tmux -2'
 fi
 
 # Display the directory structure better.
 #   -A                     : Print ANSI lines graphic indentation lines
 #   -C                     : Turn colorization on always
 #   -F                     : Appends '/', '=', '*', '@', '|' or '>' as per ls -F
-#   -h                     : Print the size in a more human readable way
+#   -h                     : Print the size in a more human-readable way
 #   --dirsfirst            : List directories before files (-U disables)
-alias tree='tree --dirsfirst -A -C -F -h'
+if [[ "${PERSONAL_tree}" == "true" ]]; then
+    alias tree='tree --dirsfirst -A -C -F -h'
+fi
 
 alias untar='tar --extract --file --verbose -z'
 
-if command -v vim >/dev/null 2>&1; then
-  alias vi='vim'
+if isPackageInstalled nvim; then
+    alias vim='nvim'
+    alias vi='nvim'
+else
+    if isPackageInstalled nvim; then
+        alias vi='vim'
+    fi
 fi
-if command -v nvim >/dev/null 2>&1; then
-  alias vim='nvim'
-  alias  vi='nvim'
+
+if [[ "${PERSONAL_wget}" == "true" ]]; then
+    alias wget='wget -c'
 fi
 
-alias wget='wget -c'
-
-
-if command -v yadm >/dev/null 2>&1; then
-    alias   y='yadm '
-	alias yad='yadm add'
-	alias ybr='yadm branch'
-	alias yco='yadm commit'
-	alias ydi='yadm diff'
-	alias ych='yadm checkout '
+# yadm = yet aother dotfile manager
+if isPackageInstalled yadm; then
+    alias y='yadm '
+    alias yad='yadm add'
+    alias ybr='yadm branch'
+    alias yco='yadm commit'
+    alias ydi='yadm diff'
+    alias ych='yadm checkout '
     alias ypu='yadm pull'
-	alias yst='yadm status'
+    alias yst='yadm status'
 fi
-
 
 mkdir ~/.local/share/Trash
 alias trash='mv --force -t ~/.local/share/Trash '
 alias emptytrash='rm ~/.local/share/Trash/*'
-
 
 echo ""
 echo "**** $HOME/.bash_aliases **** ends ****"
